Finish add, edit and delete functionality

+++ SurveyConfiguratorApp/Main.cs
@@ -2,62 +2,72 @@
 using System.Data.SqlClient;
 using System.Windows.Forms;
 using System.Configuration;
-using System.Collections.Generic;
-using System.ComponentModel;
-using SurveyConfiguratorApp.QuestionComparers;
 using SortOrder = System.Data.SqlClient.SortOrder;
 
 namespace SurveyConfiguratorApp
 {
     public partial class Main : Form
     {
-        QuestionManager questionManager;
-        OrderingMethod orderingMethod;
+        IRepository<Question> questionManager;
+        SortMethod sortMethod;
         SortOrder sortOrder;
 
         public Main()
         {
             InitializeComponent();
-            questionManager = new QuestionManager(ConfigurationManager.ConnectionStrings["surveyConnection"].ConnectionString);
-            orderingMethod = OrderingMethod.ByQuestionText;
+            sortMethod = SortMethod.ByQuestionText;
             sortOrder = SortOrder.Ascending;
+            try
+            {
+                string connectionString = ConfigurationManager.ConnectionStrings["surveyConnection"].ConnectionString;
+                if (connectionString != null)
+                    questionManager = new QuestionManager(connectionString);
+
+            }
+            catch (Exception ex)
+            {
+                showError("Invalid Connection string , Please check configuration file");
+            }
         }
 
         private void Main_Load(object sender, EventArgs e)
         {
-            refreshButton.PerformClick();
-
+            if (questionManager != null)
+                refreshButton.PerformClick();
+            else
+                Application.Exit();
         }
 
         private void questionDataGridView_ColumnHeaderMouseClick_1(object sender, DataGridViewCellMouseEventArgs e)
         {
-            //Toggle Sort
-            if (sortOrder == SortOrder.Ascending)
-                sortOrder = SortOrder.Descending;
-            else
-                sortOrder = SortOrder.Ascending;
             switch (e.ColumnIndex)
             {
                 case 0:
-                    orderingMethod = OrderingMethod.ByQuestionText;
+                    setSortMethod(SortMethod.ByQuestionText);
                     break;
                 case 1:
-                    orderingMethod = OrderingMethod.ByType;
+                    setSortMethod(SortMethod.ByType);
                     break;
                 case 2:
-                    orderingMethod = OrderingMethod.ByOrder;
+                    setSortMethod(SortMethod.ByOrder);
                     break;
             }
-            questionManager.OrderList(orderingMethod, sortOrder);
-            questionDataGridView.DataSource = questionManager.Items;
-            questionDataGridView.Refresh();
+            questionManager.OrderList(sortMethod, sortOrder);
+            refreshList();
 
         }
 
         private void refreshButton_Click(object sender, EventArgs e)
         {
-            questionManager.Refresh();
-            refreshList();
+            try
+            {
+                questionManager.Refresh();
+                refreshList();
+            }
+            catch (Exception ex)
+            {
+                showError(ex.Message);
+            }
         }
 
         private void questionDataGridView_SelectionChanged(object sender, EventArgs e)
@@ -76,18 +86,28 @@ namespace SurveyConfiguratorApp
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
-            int selectedRow = -1;
-            if (questionDataGridView.SelectedRows.Count > 0)
-                selectedRow = questionDataGridView.SelectedRows[0].Index;
-
-            if (selectedRow >= 0)
+            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete Question?", "Delete Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (dialogResult == DialogResult.Yes)
             {
-                questionManager.Delete(questionManager.Items[selectedRow].ID);
-                refreshList();
-            }
-            else
-                showError("No question is Selected");
+                int selectedRow = -1;
+                if (questionDataGridView.SelectedRows.Count > 0)
+                    selectedRow = questionDataGridView.SelectedRows[0].Index;
 
+                if (selectedRow >= 0)
+                {
+                    try
+                    {
+                        questionManager.Delete(questionManager.Items[selectedRow].ID);
+                        refreshList();
+                    }
+                    catch (Exception ex)
+                    {
+                        showError(ex.Message);
+                    }
+                }
+                else
+                    showError("No question is Selected");
+            }
         }
 
         private void addButton_Click(object sender, EventArgs e)
@@ -95,11 +115,15 @@ namespace SurveyConfiguratorApp
             QuestionProperties propertiesDialog = new QuestionProperties();
             if (propertiesDialog.ShowDialog(this) == DialogResult.OK)
             {
-
-            }
-            else
-            {
-
+                try
+                {
+                    questionManager.Insert(propertiesDialog.question);
+                    refreshList();
+                }
+                catch (Exception ex)
+                {
+                    showError(ex.Message);
+                }
             }
             propertiesDialog.Dispose();
         }
@@ -115,11 +139,15 @@ namespace SurveyConfiguratorApp
                 QuestionProperties propertiesDialog = new QuestionProperties(questionManager.Items[selectedRow]);
                 if (propertiesDialog.ShowDialog(this) == DialogResult.OK)
                 {
-
-                }
-                else
-                {
-
+                    try
+                    {
+                        questionManager.Update(propertiesDialog.question);
+                        refreshList();
+                    }
+                    catch (Exception ex)
+                    {
+                        showError(ex.Message);
+                    }
                 }
                 propertiesDialog.Dispose();
             }
@@ -129,7 +157,7 @@ namespace SurveyConfiguratorApp
 
         private void refreshList()
         {
-            if (questionManager.isConnectedToDB())
+            if (questionManager.IsConnected())
             {
                 try
                 {
@@ -154,9 +182,24 @@ namespace SurveyConfiguratorApp
             }
 
         }
-        private void showError(string errorMessage)
+        public static void showError(string errorMessage)
         {
             MessageBox.Show(errorMessage, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+        }
+
+        private void toggleSortOrder()
+        {
+            if (sortOrder == SortOrder.Ascending)
+                sortOrder = SortOrder.Descending;
+            else
+                sortOrder = SortOrder.Ascending;
+        }
+        private void setSortMethod(SortMethod newSortMethod)
+        {
+            if (newSortMethod == sortMethod)
+                toggleSortOrder();
+            else
+                sortMethod = newSortMethod;
         }
     }
 }
